requirements.txt
google_api_python_client==2.49.0
httplib2==0.20.4
oauth2client==4.1.3
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
requests==2.27.1
SQLAlchemy==1.4.31
SQLAlchemy_Utils==0.38.1

